Completed exercises 4.18 -- 4.22

+++ part4/blog_list_ts/controllers/blogs.ts
@@ -2,6 +2,9 @@ import { Router } from 'express'
 import Blog, { IBlog } from '../models/blog'
 import logger from '../utils/logger'
 import User from '../models/user'
+import jwt from 'jsonwebtoken'
+import express from 'express'
+import { BlogRequest } from '../utils/middleware'
 
 const blogsRouter = Router()
 
@@ -10,25 +13,31 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', async (request: express.Request, response) => {
   const body: IBlog = request.body
-  const user = await User.findById(body.user)
-
-  if (user) {
-    const blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes ? body.likes : 0,
-      user: user.id,
-    })
-    const savedBlog = await blog.save()
-    console.log(savedBlog)
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
-    response.status(201).json(savedBlog)
+  const req = request as BlogRequest
+  // console.log('Token at post', req.token)
+  const decodedToken = jwt.verify(req.token, process.env.SECRET as string)
+  if (typeof decodedToken == 'string' || !decodedToken.id) {
+    response.status(401).json({ error: 'token invalid' })
   } else {
-    response.status(400).end()
+    const user = await User.findById(decodedToken.id)
+    if (user) {
+      const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes ? body.likes : 0,
+        user: user.id,
+      })
+      const savedBlog = await blog.save()
+      console.log(savedBlog)
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
+      response.status(201).json(savedBlog)
+    } else {
+      response.status(400).json({ error: 'user token invalid' })
+    }
   }
 })
 
@@ -41,34 +50,58 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.put('/:id', async (request, response) => {
-  const { title, author, url, likes } = request.body
+blogsRouter.put(
+  '/:id',
+  async (request: express.Request, response: express.Response) => {
+    const updBlog = await Blog.findById(request.params.id)
+    const req = request as BlogRequest
+    console.log('Token at post', req.user, updBlog?.user?.toString())
+    if (!updBlog || !updBlog.user) {
+      response.status(401).json({ error: 'blog id invalid' })
+    } else if (req.user != updBlog.user.toString()) {
+      response.status(401).json({ error: 'token invalid' })
+    } else {
+      const { title, author, url, likes } = request.body
 
-  const updatedNote = await Blog.findByIdAndUpdate(
-    request.params.id,
-    { title, author, url, likes },
-    { new: true, runValidators: true, context: 'query' },
-  )
-  if (updatedNote) {
-    response.json(updatedNote)
-  } else {
-    response.status(404).end()
-  }
-})
-
-blogsRouter.delete('/:id', async (request, response) => {
-  const delBlog = await Blog.findByIdAndDelete(request.params.id)
-  logger.info(['deleted note', delBlog])
-  if (delBlog) {
-    const user = await User.findById(delBlog.user)
-    if (user) {
-      user.blogs = user.blogs.filter((n) => n != delBlog._id)
-      await user.save()
-      response.status(204).end()
+      const updatedNote = await Blog.findByIdAndUpdate(
+        request.params.id,
+        { title, author, url, likes },
+        { new: true, runValidators: true, context: 'query' },
+      )
+      if (updatedNote) {
+        response.json(updatedNote)
+      } else {
+        response.status(404).end()
+      }
     }
-  } else {
-    response.status(204).end()
-  }
-})
+  },
+)
+
+blogsRouter.delete(
+  '/:id',
+  async (request: express.Request, response: express.Response) => {
+    const delBlog = await Blog.findById(request.params.id)
+    const req = request as BlogRequest
+    console.log('Token at post', req.user, delBlog?.user?.toString())
+    if (!delBlog || !delBlog.user) {
+      response.status(401).json({ error: 'blog id invalid' })
+    } else if (req.user != delBlog.user.toString()) {
+      response.status(401).json({ error: 'token invalid' })
+    } else {
+      const delBlog = await Blog.findByIdAndDelete(request.params.id)
+      logger.info(['deleted note', delBlog])
+      if (delBlog) {
+        const user = await User.findById(delBlog.user)
+        if (user) {
+          user.blogs = user.blogs.filter((n) => n != delBlog._id)
+          await user.save()
+          response.status(204).end()
+        }
+      } else {
+        response.status(204).end()
+      }
+    }
+  },
+)
 
 export default blogsRouter
